Custom Collection Fixes
Added "None" option to Apply Collection Override to revert collection merge. Similar to "None" not applying any custom collection in the global merge / snapshot applying in settings.

Sorting collections alphabetically with "None" at the top to match penumbra collection order.

Fixed custom collections not applying properly when the collection is not in focus inside penumbra.

---

Collections are very weird in penumbra and this code probably needs to be revisited again in the future properly. There is a focus/unfocus issue, fundamental limitation in how penumbra handles collection states.

1. When a collection is unfocused: Penumbra marks it as "does not exist or is not loaded".
2. Internal cache becomes inaccessible: ResolvedFiles returns 0 entries even after cache recreation
3. Public API also fails: GetChangedItemsForCollection returns 0 entires
4. Collection identity becomes inaccessible: Cannot get GUID for fallback operations.

This appears to be by design in Penumbra - unfocused collections are intentionally made inaccessible to prevent conflicts?

Global Custom Collection Overrides defined in settings seems to not apply the overridden mods if you open penumbra, focus the same named / matching collection, unfocus into a different collection, then attempt to load a snapshot. You have to load the snapshot twice for it to apply the custom collection, but only if the collection was unfocused in penumbra.